+++ src/Helpers/EConnectionParams.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace ZekyWolf\LGSQ\Helpers;
+
+abstract class EConnectionParams
+{
+    public const TYPE = 'type';
+    public const IP = 'ip';
+    public const PORT = 'port';
+    public const QPORT = 'qport';
+    public const SPORT = 'sport';
+    public const STATUS = 'status';
+    public const ERROR = '_error';
+}