+++ src/Helpers/GameTypeScheme.php
@@ -7,19 +7,19 @@ class GameTypeScheme
     public static function get(string $type): string
     {
         $schemes = [
-            'bfbc2' => 'tcp',
-            'bf3' => 'tcp',
-            'discord' => 'http',
-            'eco' => 'http',
-            'farmsim' => 'http',
-            'fivem' => 'http',
-            'ragemp' => 'http',
-            'scum' => 'http',
-            'terraria' => 'http',
-            'ts' => 'tcp',
-            'ts3' => 'tcp',
-            'teaspeak' => 'tcp',
-            'wow' => 'tcp',
+            Games::BATTLEFIELDBADCOMPANY2 => 'tcp',
+            Games::BATTLEFIELD3 => 'tcp',
+            Games::DISCORD => 'http',
+            Games::ECO => 'http',
+            Games::FARMINGSIMULATOR => 'http',
+            Games::FIVEM => 'http',
+            Games::RAGEMP => 'http',
+            Games::SCUM => 'http',
+            Games::TERRARIA => 'http',
+            Games::TS => 'tcp',
+            Games::TS3 => 'tcp',
+            Games::TEASPEAK => 'tcp',
+            Games::WOW => 'tcp',
         ];
 
         return isset($schemes[$type]) ? $schemes[$type] : 'udp';