+++ src/Helpers/Games.php
@@ -1,0 +1,426 @@
+<?php
+
+namespace ZekyWolf\LGSQ\Helpers;
+
+abstract class Games
+{
+    /** Americas Army */
+    public const AMERICASARMY               = 'aarmy';
+
+    /** Americas Army 3 */
+    public const AMERICASARMY3              = 'aarmy3';
+
+    /** Arca Sim Racing */
+    public const ARCASIMRACING              = 'arcasimracing';
+
+    /** ArmA: Armed Assault */
+    public const ARMA                       = 'arma';
+
+    /** ArmA 2 */
+    public const ARMA2                      = 'arma2';
+
+    /** ArmA 3 */
+    public const ARMA3                      = 'arma3';
+
+    /** DayZ */
+    public const DAYZ                       = 'arma3';
+
+    /** Aliens VS. Predator 2 */
+    public const ALIENSVSPREDATOR2          = 'avp2';
+
+    /** Aliens VS. Predator ( 2010 By Rebellion ) */
+    public const ALIENSVSPREDATOR2_2010     = 'avp2010';
+
+    /** Battlefield Bad Company 2 */
+    public const BATTLEFIELDBADCOMPANY2     = 'bfbc2';
+
+    /** Battlefield Vietnam */
+    public const BATTLEFIELDVIETNAM         = 'bfvietnam';
+
+    /** Battlefield 1942 */
+    public const BATTLEFIELD1942            = 'bf1942';
+
+    /** Battlefield 2 */
+    public const BATTLEFIELD2               = 'bf2';
+
+    /** Battlefield 3 */
+    public const BATTLEFIELD3               = 'bf3';
+
+    /** Battlefield 4 */
+    public const BATTLEFIELD4               = 'bf4';
+
+    /** Battlefield 2142 */
+    public const BATTLEFIELD2142            = 'bf2142';
+
+    /** Call Of Duty */
+    public const CALLOFDUTY                 = 'callofduty';
+
+    /** Call Of Duty: Black Ops 3 */
+    public const CALLOFDUTYBO3              = 'callofdutybo3';
+
+    /** Call Of Duty (IW4, IW5, IW6x) */
+    public const CALLOFDUTYIW               = 'callofdutyiw';
+
+    /** Call Of Duty: United Offensive */
+    public const CALLOFDUTYOU              = 'callofdutyuo';
+
+    /** Call Of Duty: World at War */
+    public const CALLOFDUTYWAW              = 'callofdutywaw';
+
+    /** Call Of Duty 2 */
+    public const CALLOFDUTY2                = 'callofduty2';
+
+    /** Call Of Duty 4 / CoD4X */
+    public const CALLOFDUTY4                = 'callofduty4';
+
+    /** Command and Conquer: Renegad */
+    public const COMMANDANDCONQUER          = 'cncrenegade';
+
+    /** Conan Exiles */
+    public const CONANEXILES                = 'conanexiles';
+
+    /** Crysis */
+    public const CRYSIS                     = 'crysis';
+
+    /** Crysis Wars */
+    public const CRYSISWARS                 = 'crysiswars';
+
+    /** Counter-Strike 2D */
+    public const CS2D                       = 'cs2d';
+
+    /** Cube Engine */
+    public const CUBE                       = 'cube';
+
+    /** Discord */
+    public const DISCORD                    = 'discord';
+
+    /** Doom - Skulltag */
+    public const DOOMSSKULLTAG              = 'doomskulltag';
+
+    /** Doom - ZDaemon */
+    public const DOOMZDEAMON                = 'doomzdaemon';
+
+    /** Doom 3 */
+    public const DOOM3                      = 'DOOM3';
+
+    /** Deer Hunter 2005 */
+    public const DEERHUNTER2005             = 'dh2005';
+
+    /** ECO */
+    public const ECO                        = 'eco';
+
+    /** Factorio */
+    public const FACTORIO                   = 'factorio';
+
+    /** Far Cry */
+    public const FARCRY                     = 'farcry';
+
+    /** Farming Simulator */
+    public const FARMINGSIMULATOR           = 'farmsim';
+
+    /** F.E.A.R. */
+    public const FEAR                       = 'fear';
+
+    /** FiveM */
+    public const FIVEM                      = 'fivem';
+
+    /** RedM */
+    public const REDM                       = 'fivem';
+
+    /** Operation Flashpoint */
+    public const OPERATIONFLASHPOINT        = 'flashpoint';
+
+    /** Freelancer */
+    public const FREELANCER                 = 'freelancer';
+
+    /** Frontlines: Fuel Of War */
+    public const FRONTLINES                 = 'frontlines';
+
+    /** F1 Challenge 99-02 */
+    public const F1CHALLENGE9902            = 'f1c9902';
+
+    /** Generic GameSpy 1 */
+    public const GENERICSPY1                = 'gamespy1';
+
+    /** Generic GameSpy 2 */
+    public const GENERICSPY2                = 'gamespy2';
+
+    /** Generic GameSpy 3 */
+    public const GENERICSPY3                = 'gamespy3';
+
+    /** Ghost Recon */
+    public const GHOSTRECON                 = 'ghostrecon';
+
+    /** Ghost Recon: Advanced Warfighter */
+    public const GRAW                       = 'graw';
+
+    /** Ghost Recon: Advanced Warfighter 2 */
+    public const GRAW2                      = 'graw2';
+
+    /** GTR 2 */
+    public const GTR2                       = 'gtr2';
+
+    /** Hidden and Dangerous 2 */
+    public const HIDDENANDDANGEROUSE2       = 'had2';
+
+    /** Half-life Steam Protocol (CS 1.6, etc) */
+    public const HALFLIFE                   = 'halflife';
+
+    /** Half-Life WON Protocol [OLD] (CS 1.5) */
+    public const HALFLIFEWON                = 'halflifewon';
+
+    /** Halo */
+    public const HALO                       = 'halo';
+
+    /** IL-2 Sturmovik */
+    public const IL22STURMOVIK              = 'il2';
+
+    /** JediKnight 2: Jedi Outcast */
+    public const JEDIKNIGHT2                = 'jediknight2';
+
+    /** JediKnight: Jedi Academy */
+    public const JEDIKNIGHTJA               = 'jediknightja';
+
+    /** Just Cause 2 Multiplayer */
+    public const JUSTCOUSE2MP               = 'jc2mp';
+
+    /** Killing Floor */
+    public const KILLINGFLOOR               = 'killingfloor';
+
+    /** Kingpin: Life of Crime */
+    public const KINGPIN                    = 'kingpin';
+
+    /** Mafia II Multiplayer */
+    public const MAFIA2MP                   = 'm2mp';
+
+    /** Minecraft */
+    public const MINECRAFT                  = 'minecraft';
+
+    /** Medal of Honor: Allied Assault */
+    public const MEDALOFHONOR_ALLIEDASSAULT = 'mohaa';
+
+    /** Medal of Honor: Allied Assault */
+    public const MOHAA                      = 'mohaa';
+
+    /** Medal of Honor: Allied Assault Breakthrough */
+    public const MEDALOFHONOR_ALLIEDASSAULTBREAKTHROUGH = 'mohaab';
+
+    /** Medal of Honor: Allied Assault Breakthrough */
+    public const MOHAAB                     = 'mohaab';
+
+    /** Medal of Honor: Allied Assault Spearhead */
+    public const MEDALOFHONOR_ALLIEDASSAULTSPEARHEAD = 'mohaas';
+
+    /** Medal of Honor: Allied Assault Spearhead */
+    public const MOHAAS                      = 'mohaas';
+
+    /** Medal of Honor: Pacific Assault */
+    public const MEDALOFHONOR_PACIFICASSAULT  = 'mohpa';
+
+    /** Medal of Honor: Pacific Assault */
+    public const MOPHA                      = 'mohpa';
+
+    /** Multi Theft Auto*/
+    public const MULTITHEFTAUTO             = 'mta';
+
+    /** Multi Theft Auto */
+    public const MTA                        = 'mta';
+
+    /** Mumble */
+    public const MUMBLE                     = 'mumble';
+
+    /** Nascar Thunder 2004 */
+    public const NASCAR2004                 = 'nascar2004';
+
+    /** NeverWinter Nights */
+    public const NEVERWINTER                = 'neverwinter';
+
+    /** NeverWinter Nights 2 */
+    public const NEVERWINTER2               = 'neverwinter2';
+
+    /** Nexuiz */
+    public const NEXUIZ                     = 'nexuiz';
+
+    /** Open Transport Tycoon Deluxe */
+    public const OPENTTD                    = 'openttd';
+
+    /** PainKiller */
+    public const PAINKILLER                 = 'painkiller';
+
+    /** Plain Sight */
+    public const PLAINSIGHT                 = 'plainsight';
+
+    /** Prey */
+    public const PREY                       = 'prey';
+
+    /** Quake World */
+    public const QUAKEWORLD                 = 'quakeworld';
+
+    /** Enemy Territory: Quake Wars */
+    public const QUAKEWARS                  = 'quakewars';
+
+    /** Quake 2 */
+    public const QUAKE2                     = 'quake2';
+
+    /** Quake 3 */
+    public const QUAKE3                     = 'quake3';
+
+    /** Quake 4 */
+    public const QUAKE4                     = 'quake4';
+
+    /** Rage:MP */
+    public const RAGEMP                     = 'ragemp';
+
+    /** Raven Shield */
+    public const RAVENSHIELD                = 'ravenshield';
+
+    /** Red Orchestra */
+    public const REDORCHESTRA               = 'redorchestra';
+
+    /** RFactor */
+    public const RFACTOR                    = 'rfactor';
+
+    /** San Andreas Multiplayer */
+    public const SAMP                       = 'samp';
+
+    /** Savage */
+    public const SAVAGE                     = 'savage';
+
+    /** Savage 2 */
+    public const SAVAGE2                    = 'savage2';
+
+    /** SCUM */
+    public const SCUM                       = 'scum';
+
+    /** Satisfactory */
+    public const SF                         = 'sf';
+
+    /** Satisfactory */
+    public const SATISFACTORY               = 'sf';
+
+    /** Serious Sam */
+    public const SERIOUSSAM                 = 'serioussam';
+
+    /** Serious Sam 2 */
+    public const SERIOUSSAM2                = 'serioussam2';
+
+    /** Shattered Horizon */
+    public const SHATTEREDHORIZON           = 'shatteredh';
+
+    /** Soldier of Fortune 2 */
+    public const SOLDIEROFFORTUNE2          = 'sof2';
+
+    /** Soldat */
+    public const SOLDAT                     = 'soldat';
+
+    /** Source Protocol ( Half-Life 2, etc. ) */
+    public const SOURCE                     = 'source';
+
+    /** */
+    //  public const SOURCE2                    = 'source2';
+
+    /** S.T.A.L.K.E.R. */
+    public const STALKER                    = 'stalker';
+
+    /** S.T.A.L.K.E.R. Call of Pripyat */
+    public const STALKERCOP                 = 'stalkercop';
+
+    /** S.T.A.L.K.E.R. Clear Sky */
+    public const STALKERCS                  = 'stalkercs';
+
+    /** StarTrek Elite-Force */
+    public const STARTREKEF                 = 'startrekef';
+
+    /** Star Wars: Battlefront */
+    public const STARWARSBF                 = 'starwarsbf';
+
+    /** Star Wars: Battlefront 2 */
+    public const STARWARSBF2                = 'starwarsbf2';
+
+    /** Star Wars: Republic Commando */
+    public const STARWARSRC                 = 'starwarsrc';
+
+    /** SWAT 4 */
+    public const SWAT4                      = 'swat4';
+
+    /** Teeworlds */
+    public const TEENWORLDS                 = 'teeworlds';
+
+    /** Terraria */
+    public const TERRARIA                   = 'terraria';
+
+    /** Tribes ( Starsiege ) */
+    public const TRIBERS                    = 'tribes';
+
+    /** Tribes 2 */
+    public const TRIBERS2                   = 'tribes2';
+
+    /** Tribes Vengeance */
+    public const TRIBERSV                   = 'tribesv';
+
+    /** Teamspeak */
+    public const TS                         = 'ts';
+
+    /** Teamspeak */
+    public const TEAMSPEAK                  = 'ts';
+
+    /** Teamspeak 3*/
+    public const TS3                        = 'ts3';
+
+    /** Teamspeak 3*/
+    public const TEAMSPEAK3                 = 'ts3';
+
+    /** Teaspeak */
+    public const TEASPEAK                   = 'teaspeak';
+
+    /** UrbanTerror */
+    public const URBANTERROR                = 'urbanterror';
+
+    /** Unreal Tournament */
+    public const UNREALTOURNAMENT           = 'ut';
+
+    /** Unreal Tournament 2003*/
+    public const UNREALTOURNAMENT2003       = 'ut2003';
+
+    /** Unreal Tournament 2004*/
+    public const UNREALTOURNAMENT2004       = 'ut2004';
+
+    /** Unreal Tournament 3*/
+    public const UNREALTOURNAMENT3          = 'ut3';
+
+    /** Vice City Multiplayer */
+    public const VICECITYMP                 = 'vcmp';
+
+    /** Vice City Multiplayer */
+    public const VCMP                       = 'vcmp';
+
+    /** Vietcong */
+    public const VIETCONG                   = 'vietcong';
+
+    /** Vietcong 2*/
+    public const VIETCONG2                  = 'vietcong2';
+
+    /** Warsow */
+    public const WARSOW                     = 'warsow';
+
+    /** Warsow ( 0.4.2 and older ) */
+    public const WARSOWOLD                  = 'warsowold';
+
+    /** Wolfenstein: Enemy Territory */
+    public const WOLFENSTEINET              = 'wolfet';
+
+    /** Wolfenstein: Return To Castle Wolfenstein */
+    public const WOLFENSTEINFRTCW           = 'wolfrtcw';
+
+    /** Wolfenstein ( 2009 By Raven ) */
+    public const WOLFENSTEIN2009            = 'wolf2009';
+
+    /** World of Warcraft */
+    public const WOW                        = 'wow';
+
+    /** World of Warcraft */
+    public const WORLDOFWARCRAFT            = 'wow';
+
+    /** ~ Test ( For PHP Developers ) */
+    public const TEST                       = 'test';
+}